implementação da classe Save em ComputerRepository
