+++ Repositories/ComputerRepository.cs
@@ -38,4 +38,19 @@ class ComputerRepository
 
         return computers;
     }
+
+    public void Save(Computer computer)
+    {
+        var connection = new SqliteConnection(_databaseConfig.ConnectionString);
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processador)";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processador", computer.Processor);
+
+        command.ExecuteNonQuery();
+        connection.Close();
+    }
 }
